draft for modular structure of SpectrumFitter
